search and install packs
