+++ Tests/Marketplace/Tests/search_and_install_packs_test.py
@@ -82,13 +82,13 @@ MOCK_PACKS_INSTALLATION_RESULT = [
 PACKS_PACK_META_FILE_NAME = 'pack_metadata.json'
 
 
-def mocked_generic_request_func(self, path: str, method, body=None, accept=None, _request_timeout=None, response_type='object'):
+def mocked_generic_request_func(self, path: str, method, body=None, **kwargs):
     if body:
         if body[0].get('id') == 'HelloWorld':
             return MOCK_HELLOWORLD_SEARCH_RESULTS, 200, None
         elif body and body[0].get('id') == 'AzureSentinel':
             return MOCK_AZURESENTINEL_SEARCH_RESULTS, 200, None
-    return None, None, None
+    raise ApiException(status=400)
 
 
 class MockConfiguration:
@@ -141,6 +141,9 @@ def test_search_and_install_packs_and_their_dependencies(mocker, use_multithread
 
     client = MockClient()
 
+    import time
+    mocker.patch.object(time, 'sleep')
+
     mocker.patch.object(script, 'install_packs')
     mocker.patch.object(demisto_client, 'generic_request_func', side_effect=mocked_generic_request_func)
     mocker.patch.object(script, 'is_pack_deprecated', return_value=False)  # Relevant only for post-update unit-tests
@@ -178,7 +181,8 @@ def test_search_and_install_packs_and_their_dependencies_with_error(mocker, erro
         Ensure the function returns a 'success' value of 'False'.
     """
     client = MockClient()
-
+    import time
+    mocker.patch.object(time, 'sleep')
     mocker.patch.object(script, 'install_packs')
     mocker.patch.object(script, 'fetch_pack_metadata_from_gitlab', return_value={"hidden": False})
     mocker.patch.object(demisto_client, 'generic_request_func', side_effect=ApiException(status=error_code))
@@ -406,7 +410,8 @@ class TestInstallPacks:
         http_resp = MockHttpRequest(GCP_TIMEOUT_EXCEPTION_RESPONSE_BODY)
         mocker.patch.object(demisto_client, 'generic_request_func', side_effect=ApiException(http_resp=http_resp))
         client = MockClient()
-        assert not script.install_packs(client, 'my_host', packs_to_install=[{'id': 'pack1'}, {'id': 'pack3'}])
+        success, _ = script.install_packs(client, 'my_host', packs_to_install=[{'id': 'pack1'}, {'id': 'pack3'}])
+        assert not success
 
     def test_malformed_pack_exception(self, mocker):
         """
@@ -422,7 +427,8 @@ class TestInstallPacks:
         http_resp = MockHttpRequest(MALFORMED_PACK_RESPONSE_BODY)
         mocker.patch.object(demisto_client, 'generic_request_func', side_effect=ApiException(http_resp=http_resp))
         client = MockClient()
-        assert not script.install_packs(client, 'my_host', packs_to_install=[{'id': 'pack1'}, {'id': 'pack2'}])
+        success, _ = script.install_packs(client, 'my_host', packs_to_install=[{'id': 'pack1'}, {'id': 'pack2'}])
+        assert not success
 
 
 def test_malformed_pack_id():
