+++ Tests/Marketplace/search_and_install_packs.py
@@ -1,27 +1,26 @@
 
-import contextlib
-import traceback
-from functools import lru_cache
 import base64
+import contextlib
 import glob
 import json
 import os
-import time
 import re
 import sys
-from concurrent.futures import ThreadPoolExecutor
+import traceback
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from functools import lru_cache
 from pathlib import Path
-from requests import Session
 from threading import Lock
 
 import demisto_client
-from demisto_client.demisto_api.rest import ApiException
 from demisto_sdk.commands.common import tools
 from demisto_sdk.commands.content_graph.common import PACK_METADATA_FILENAME
 from google.cloud.storage import Bucket  # noqa
 from packaging.version import Version
-from urllib3.exceptions import HTTPWarning, HTTPError
+from requests import Session
 
+from Tests.Marketplace.common import ALREADY_IN_PROGRESS
+from Tests.Marketplace.common import wait_until_not_updating, generic_request_with_retries
 from Tests.Marketplace.marketplace_constants import (IGNORED_FILES,
                                                      PACKS_FOLDER,
                                                      PACKS_FULL_PATH,
@@ -33,7 +32,6 @@ from Tests.scripts.utils import logging_wrapper as logging
 
 PACK_PATH_VERSION_REGEX = re.compile(fr'^{GCPConfig.PRODUCTION_STORAGE_BASE_PATH}/[A-Za-z0-9-_.]+/(\d+\.\d+\.\d+)/[A-Za-z0-9-_.]'  # noqa: E501
                                      r'+\.zip$')
-SUCCESS_FLAG = True
 WLM_TASK_FAILED_ERROR_CODE = 101704
 
 GITLAB_SESSION = Session()
@@ -198,7 +196,6 @@ def create_dependencies_data_structure(response_data: dict, dependants_ids: list
 
 def get_pack_dependencies(client: demisto_client,
                           pack_id: str,
-                          lock: Lock,
                           attempts_count: int = 5,
                           sleep_interval: int = 60,
                           ) -> dict | None:
@@ -208,57 +205,36 @@ def get_pack_dependencies(client: demisto_client,
     Args:
         client (demisto_client): The configured client to use.
         pack_id (str): ID of the pack to get dependencies for.
-        lock (Lock): A lock object.
         attempts_count (int): The number of attempts to install the packs.
         sleep_interval (int): The sleep interval, in seconds, between request retry attempts.
 
     Returns:
         dict | None: API response data for the /search/dependencies endpoint. None if the request failed.
     """
-    global SUCCESS_FLAG
-
     api_endpoint = "/contentpacks/marketplace/search/dependencies"
     body = [{"id": pack_id}]  # Not specifying a "version" key will return the latest version of the pack.
 
-    for attempt in range(attempts_count - 1, -1, -1):
-        logging.debug(f"Fetching dependencies information for '{pack_id}' using Marketplace API "
-                      f"(Attempt {attempts_count - attempt}/{attempts_count})")
-        try:
-            response_data = demisto_client.generic_request_func(
-                client,
-                path=api_endpoint,
-                method='POST',
-                body=body,
-                accept='application/json',
-                _request_timeout=None,
-                response_type='object',
-            )[0]
+    def success_handler(response):
+        logging.success(f"Succeeded to fetch dependencies for pack '{pack_id}'")
+        return True, response
 
-            logging.debug(f"Succeeded to fetch dependencies for pack '{pack_id}'.\nResponse: '{json.dumps(response_data)}'")
-            return response_data
+    failure_massage = f"Failed to fetch dependencies for pack: {pack_id}"
+    retries_message = f"Retrying to fetch dependencies for pack: {pack_id}"
+    prior_message = f"Fetching dependencies information for pack: {pack_id} using Marketplace API"
 
-        except Exception as ex:
-            if isinstance(ex, ApiException):
-                logging.warning(f"API request to fetch dependencies of pack '{pack_id}' has failed.\n"
-                                f"Response code '{ex.status}'\nResponse: '{ex.body}'\nResponse Headers: '{ex.headers}'")
-
-            elif isinstance(ex, HTTPError | HTTPWarning):
-                logging.warning(f"Failed to perform HTTP request : {ex}")
-
-            else:
-                logging.warning(f"API call to fetch dependencies of '{pack_id}' has failed.\nError: {ex}.")
-                logging.debug("Stack trace:\n" + traceback.format_exc())
-
-            if attempt:  # There are remaining retry attempts
-                logging.debug(f"Sleeping for {sleep_interval} seconds and retrying...")
-                time.sleep(sleep_interval)
-
-            else:  # No retry attempts remaining
-                logging.error(f"All {attempts_count} attempts have failed.")
-                with lock:
-                    SUCCESS_FLAG = False
-
-    return None
+    _, data = generic_request_with_retries(client=client,
+                                           retries_message=retries_message,
+                                           exception_message=failure_massage,
+                                           prior_message=prior_message,
+                                           path=api_endpoint,
+                                           method='POST',
+                                           response_type='object',
+                                           body=body,
+                                           request_timeout=None,
+                                           attempts_count=attempts_count,
+                                           sleep_interval=sleep_interval,
+                                           success_handler=success_handler)
+    return data
 
 
 def find_malformed_pack_id(body: str) -> list:
@@ -309,7 +285,10 @@ def handle_malformed_pack_ids(malformed_pack_ids, packs_to_install):
                             f'though it was not in the installation list')
 
 
-def install_packs_from_artifacts(client: demisto_client, host: str, test_pack_path: str, pack_ids_to_install: list):
+def install_packs_from_artifacts(client: demisto_client,
+                                 host: str,
+                                 test_pack_path: str,
+                                 pack_ids_to_install: list) -> bool:
     """
     Installs all the packs located in the artifacts folder of the BitHub actions build. Please note:
     The server always returns a 200 status even if the pack was not installed.
@@ -329,12 +308,13 @@ def install_packs_from_artifacts(client: demisto_client, host: str, test_pack_pa
         if any(pack_id in local_pack for pack_id in pack_ids_to_install):
             logging.info(f'Installing the following pack: {local_pack}')
             upload_zipped_packs(client=client, host=host, pack_path=local_pack)
+    return True
 
 
 def install_packs_private(client: demisto_client,
                           host: str,
                           pack_ids_to_install: list,
-                          test_pack_path: str):
+                          test_pack_path: str) -> bool:
     """ Make a packs installation request.
 
     Args:
@@ -343,10 +323,10 @@ def install_packs_private(client: demisto_client,
         pack_ids_to_install (list): List of Pack IDs to install.
         test_pack_path (str): Path where test packs are located.
     """
-    install_packs_from_artifacts(client,
-                                 host,
-                                 pack_ids_to_install=pack_ids_to_install,
-                                 test_pack_path=test_pack_path)
+    return install_packs_from_artifacts(client,
+                                        host,
+                                        pack_ids_to_install=pack_ids_to_install,
+                                        test_pack_path=test_pack_path)
 
 
 def get_error_ids(body: str) -> dict[int, str]:
@@ -359,7 +339,6 @@ def get_error_ids(body: str) -> dict[int, str]:
 def install_packs(client: demisto_client,
                   host: str,
                   packs_to_install: list,
-                  request_timeout: int = 3600,
                   attempts_count: int = 5,
                   sleep_interval: int = 60,
                   ):
@@ -373,87 +352,77 @@ def install_packs(client: demisto_client,
         client (demisto_client): The configured client to use.
         host (str): The server URL.
         packs_to_install (list): A list of the packs to install.
-        request_timeout (int): Timeout setting, in seconds, for the installation request.
         attempts_count (int): The number of attempts to install the packs.
         sleep_interval (int): The sleep interval, in seconds, between install attempts.
     """
-    global SUCCESS_FLAG
     if not packs_to_install:
         logging.info("There are no packs to install on servers. Consolidating installation as success")
-        return SUCCESS_FLAG
-    try:
-        for attempt in range(attempts_count - 1, -1, -1):
-            try:
-                logging.info(f"Installing packs {', '.join([p.get('id') for p in packs_to_install])} on server {host}. "
-                             f"Attempt: {attempts_count - attempt}/{attempts_count}")
-                response, status_code, headers = demisto_client.generic_request_func(client,
-                                                                                     path='/contentpacks/marketplace/install',
-                                                                                     method='POST',
-                                                                                     body={'packs': packs_to_install,
-                                                                                           'ignoreWarnings': True},
-                                                                                     accept='application/json',
-                                                                                     _request_timeout=request_timeout,
-                                                                                     response_type='object')
+        return True, []
 
-                if 200 <= status_code < 300 and status_code != 204:
-                    packs_data = [{'ID': pack.get('id'), 'CurrentVersion': pack.get('currentVersion')} for pack in response]
-                    logging.success(f'Packs were successfully installed on server {host}')
-                    logging.debug(f'The packs that were successfully installed on server {host}:\n{packs_data}')
-                    break
+    success = True
 
-                if not attempt:
-                    raise Exception(f"Got bad status code: {status_code}, headers: {headers}")
+    def success_handler(response_data):
+        packs_data = [{'ID': pack.get('id'), 'CurrentVersion': pack.get('currentVersion')} for pack in response_data]
+        logging.success('Packs ware successfully uninstalled from the server')
 
-                logging.warning(f"Got bad status code: {status_code} from the server, headers:{headers}")
+        return success, packs_data
 
-            except ApiException as ex:
-                if malformed_ids := find_malformed_pack_id(ex.body):
-                    handle_malformed_pack_ids(malformed_ids, packs_to_install)
-                    if not attempt:
-                        raise Exception(f"malformed packs: {malformed_ids}") from ex
+    def api_exception_handler(ex, attempt_left):
+        nonlocal packs_to_install, success
+        if ALREADY_IN_PROGRESS in ex.body:
+            wait_succeeded = wait_until_not_updating(client)
+            if not wait_succeeded:
+                raise Exception(
+                    "Failed to wait for the server to exit installation/updating status"
+                ) from ex
+        if malformed_ids := find_malformed_pack_id(ex.body):
+            handle_malformed_pack_ids(malformed_ids, packs_to_install)
+            if not attempt_left:
+                raise Exception(f"malformed packs: {malformed_ids}") from ex
 
-                    # We've more attempts, retrying without tho malformed packs.
-                    SUCCESS_FLAG = False
-                    logging.error(f"Unable to install malformed packs: {malformed_ids}, retrying without them.")
-                    packs_to_install = [pack for pack in packs_to_install if pack['id'] not in malformed_ids]
+            # We've more attempts, retrying without tho malformed packs.
+            success = False
+            logging.error(f"Unable to install malformed packs: {malformed_ids}, retrying without them.")
+            packs_to_install = [pack for pack in packs_to_install if pack['id'] not in malformed_ids]
 
-                error_ids = get_error_ids(ex.body)
-                if WLM_TASK_FAILED_ERROR_CODE in error_ids:
-                    if "polling request failed for task ID" in error_ids[WLM_TASK_FAILED_ERROR_CODE].lower():
-                        logging.error(f"Got {WLM_TASK_FAILED_ERROR_CODE} error code - polling request failed for task ID, "
-                                      f"retrying.")
-                    else:
-                        # If we got this error code, it means that the modeling rules are not valid, exiting install flow.
-                        raise Exception(f"Got [{WLM_TASK_FAILED_ERROR_CODE}] error code - Modeling rules and Dataset validations "
-                                        f"failed. Please look at GCP logs to understand why it failed.") from ex
+        error_ids = get_error_ids(ex.body)
+        if WLM_TASK_FAILED_ERROR_CODE in error_ids:
+            if "polling request failed for task ID" in error_ids[WLM_TASK_FAILED_ERROR_CODE].lower():
+                logging.error(f"Got {WLM_TASK_FAILED_ERROR_CODE} error code - polling request failed for task ID, "
+                              f"retrying.")
+            else:
+                # If we got this error code, it means that the modeling rules are not valid, exiting install flow.
+                raise Exception(f"Got [{WLM_TASK_FAILED_ERROR_CODE}] error code - Modeling rules and Dataset validations "
+                                f"failed. Please look at GCP logs to understand why it failed.") from ex
 
-                if not attempt:  # exhausted all attempts, understand what happened and exit.
-                    if 'timeout awaiting response' in ex.body:
-                        if '/packs/' in ex.body:
-                            pack_id = get_pack_id_from_error_with_gcp_path(ex.body)
-                            raise Exception(f"timeout awaiting response headers while trying to install pack {pack_id}") from ex
+        if not attempt_left:  # exhausted all attempts, understand what happened and exit.
+            if 'timeout awaiting response' in ex.body:
+                if '/packs/' in ex.body:
+                    pack_id = get_pack_id_from_error_with_gcp_path(ex.body)
+                    raise Exception(f"timeout awaiting response headers while trying to install pack {pack_id}") from ex
 
-                        raise Exception("timeout awaiting response headers while trying to install, "
-                                        "couldn't determine pack id.") from ex
+                raise Exception("timeout awaiting response headers while trying to install, "
+                                "couldn't determine pack id.") from ex
 
-                    if 'Item not found' in ex.body:
-                        raise Exception(f'Item not found error, headers:{ex.headers}.') from ex
+            if 'Item not found' in ex.body:
+                raise Exception(f'Item not found error, headers:{ex.headers}.') from ex
 
-                    # Unknown exception reason, re-raise.
-                    raise Exception(f"Got {ex.status} from server, message:{ex.body}, headers:{ex.headers}") from ex
-            except (HTTPError, HTTPWarning) as http_ex:
-                if not attempt:
-                    raise Exception("Failed to perform http request to the server") from http_ex
-
-            # There are more attempts available, sleep and retry.
-            logging.debug(f"Failed to install packs: {packs_to_install}. Sleeping for {sleep_interval} seconds.")
-            time.sleep(sleep_interval)
-    except Exception as e:
-        logging.exception(f'The request to install packs: {packs_to_install} has failed. Additional info: {str(e)}')
-        SUCCESS_FLAG = False
-
-    finally:
-        return SUCCESS_FLAG
+    retries_message = f"Retrying to install packs on server {host}"
+    exception_massage = f"Failed to install packs on server {host}"
+    prior_message = f"Installing packs on server {host}."
+    return generic_request_with_retries(client=client,
+                                        retries_message=retries_message,
+                                        exception_message=exception_massage,
+                                        prior_message=prior_message,
+                                        path='/contentpacks/marketplace/install',
+                                        body={'packs': packs_to_install,
+                                              'ignoreWarnings': True},
+                                        response_type='object',
+                                        method='DELETE',
+                                        attempts_count=attempts_count,
+                                        sleep_interval=sleep_interval,
+                                        success_handler=success_handler,
+                                        api_exception_handler=api_exception_handler)
 
 
 def search_pack_and_its_dependencies(client: demisto_client,
@@ -466,7 +435,7 @@ def search_pack_and_its_dependencies(client: demisto_client,
                                      commit_hash: str,
                                      multithreading: bool = True,
                                      batch_packs_install_request_body: list | None = None,
-                                     ):
+                                     ) -> bool:
     """
     Update 'packs_to_install' (a pointer to a list that's reused and updated by the function on every iteration)
     with 'pack_id' and its dependencies, if 'pack_id' is not deprecated.
@@ -490,16 +459,16 @@ def search_pack_and_its_dependencies(client: demisto_client,
             If false - install all packs in one batch.
         batch_packs_install_request_body (list | None, None): A list of pack batches (lists) to use in installation requests.
             Each list contain one pack and its dependencies.
+    # Returns: True if we succeeded to get the dependencies, False otherwise.
     """
-    global SUCCESS_FLAG
     if is_pack_deprecated(pack_id=pack_id, production_bucket=production_bucket, commit_hash=commit_hash):
         logging.warning(f"Pack '{pack_id}' is deprecated (hidden) and will not be installed.")
-        return
+        return True
 
-    api_data = get_pack_dependencies(client=client, pack_id=pack_id, lock=lock)
+    api_data = get_pack_dependencies(client=client, pack_id=pack_id)
 
     if not api_data:
-        return
+        return False
 
     dependencies_data: list[dict] = []
 
@@ -514,8 +483,7 @@ def search_pack_and_its_dependencies(client: demisto_client,
 
     except Exception as ex:
         logging.error(f"Error: {ex}\n\nStack trace:\n{traceback.format_exc()}")
-        SUCCESS_FLAG = False
-        return
+        return False
 
     if dependencies_data:
         dependencies_ids = [dependency['id'] for dependency in dependencies_data]
@@ -528,11 +496,8 @@ def search_pack_and_its_dependencies(client: demisto_client,
 
             if is_deprecated:
                 logging.critical(f"Pack '{pack_id}' depends on pack '{dependency_id}' which is a deprecated pack.")
-                SUCCESS_FLAG = False
-                return
-
-            else:
-                current_packs_to_install.append(dependency)
+                return False
+            current_packs_to_install.append(dependency)
 
     with lock:
         if not multithreading:
@@ -559,6 +524,7 @@ def search_pack_and_its_dependencies(client: demisto_client,
                     installation_request_body.append(
                         get_pack_installation_request_data(pack_id=pack['id'],
                                                            pack_version=pack['extras']['pack']['currentVersion']))
+    return True
 
 
 def get_latest_version_from_bucket(pack_id: str, production_bucket: Bucket) -> str:
@@ -633,7 +599,8 @@ def install_all_content_packs_for_nightly(client: demisto_client, host: str, ser
             pack_version = get_latest_version_from_bucket(pack_id, production_bucket)
             if pack_version:
                 all_packs.append(get_pack_installation_request_data(pack_id, pack_version))
-    install_packs(client, host, all_packs)
+    success, _ = install_packs(client, host, all_packs)
+    return success
 
 
 def install_all_content_packs_from_build_bucket(client: demisto_client, host: str, server_version: str,
@@ -711,14 +678,14 @@ def upload_zipped_packs(client: demisto_client,
         else:
             message = response_data.get('message', '')
             raise Exception(f'Failed to install packs - with status code {status_code}\n{message}')
-    except Exception:
+    except Exception:  # noqa
         logging.exception('The request to install packs has failed.')
         sys.exit(1)
 
 
 def search_and_install_packs_and_their_dependencies_private(test_pack_path: str,
                                                             pack_ids: list,
-                                                            client: demisto_client):
+                                                            client: demisto_client) -> bool:
     """ Searches for the packs from the specified list, searches their dependencies, and then installs them.
     Args:
         test_pack_path (str): Path of where the test packs are located.
@@ -733,9 +700,7 @@ def search_and_install_packs_and_their_dependencies_private(test_pack_path: str,
 
     logging.info(f'Starting to search and install packs in server: {host}')
 
-    install_packs_private(client, host, pack_ids, test_pack_path)
-
-    return SUCCESS_FLAG
+    return install_packs_private(client, host, pack_ids, test_pack_path)
 
 
 def search_and_install_packs_and_their_dependencies(pack_ids: list,
@@ -782,17 +747,30 @@ def search_and_install_packs_and_their_dependencies(pack_ids: list,
         'commit_hash': master_commit_hash,
     }
 
+    success = True
     if not multithreading:
         for pack_id in pack_ids:
-            search_pack_and_its_dependencies(pack_id=pack_id, **kwargs)
+            success &= search_pack_and_its_dependencies(pack_id=pack_id, **kwargs)
 
     else:
         with ThreadPoolExecutor(max_workers=130) as pool:
-            for pack_id in pack_ids:
-                pool.submit(search_pack_and_its_dependencies, pack_id=pack_id, **kwargs)
+            futures = [
+                pool.submit(
+                    search_pack_and_its_dependencies, pack_id=pack_id, **kwargs
+                )
+                for pack_id in pack_ids
+            ]
+            for future in as_completed(futures):
+                try:
+                    success &= future.result()
+                except Exception:  # noqa E722
+                    logging.exception(
+                        f"An exception occurred while searching for dependencies of pack '{pack_ids[futures.index(future)]}'"
+                    )
+                    success = False
         batch_packs_install_request_body = [installation_request_body]
 
     for packs_to_install_body in batch_packs_install_request_body:
         install_packs(client, host, packs_to_install_body)
 
-    return packs_to_install, SUCCESS_FLAG
+    return packs_to_install, success
